// dx/frontend/src/api/model/edgeOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * dx API
 * OpenAPI spec version: 0.1.0
 */

export interface EdgeOut {
  created: string;
  /** The source has changed since this edge was recorded */
  is_stale: boolean;
  source_id: string;
  source_version: number;
  target_id: string;
}